+++ src/acconeer/exptool/a121/_core/entities/__init__.py
@@ -20,6 +20,7 @@ from .containers import (
     Record,
     Result,
     ResultContext,
+    SensorCalibration,
     SensorInfo,
     ServerInfo,
     StackedResults,