+++ src/acconeer/exptool/a121/_core/entities/containers/__init__.py
@@ -5,6 +5,7 @@ from .client_info import ClientInfo
 from .metadata import Metadata
 from .record import PersistentRecord, Record
 from .result import Result, ResultContext
+from .sensor_calibration import SensorCalibration
 from .server_info import SensorInfo, ServerInfo
 from .stacked_results import StackedResults
 from .utils import (