+++ src/acconeer/exptool/a121/_core/mediators/recorder.py
@@ -3,7 +3,7 @@
 
 from __future__ import annotations
 
-from typing import Any
+from typing import Any, Optional
 
 from typing_extensions import Protocol
 
@@ -11,6 +11,7 @@ from acconeer.exptool.a121._core.entities import (
     ClientInfo,
     Metadata,
     Result,
+    SensorCalibration,
     ServerInfo,
     SessionConfig,
 )
@@ -24,6 +25,8 @@ class Recorder(Protocol):
         extended_metadata: list[dict[int, Metadata]],
         server_info: ServerInfo,
         session_config: SessionConfig,
+        calibrations: Optional[dict[int, SensorCalibration]],
+        calibrations_provided: dict[int, bool],
     ) -> None:
         ...