+++ src/acconeer/exptool/a121/_core/peripherals/communication/exploration_protocol/__init__.py
@@ -3,4 +3,5 @@
 
 from ._factory import get_exploration_protocol
 from ._latest import ExplorationProtocol, ExplorationProtocolError
+from ._v0_4_1 import ExplorationProtocol_0_4_1
 from .server_error import ServerError